requirements.txt
streamlit
transformers
torch
yfinance
pandas
numpy
requests
python-dotenv
feedparser
plotly
pyngrok
